create global enums
