+++ Assets/Scripts/Global.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5190e4fbb95a60b47bdfd9ccd460c009
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
