+++ Assets/Scripts/Steam/AchievementUnlocker.cs
@@ -7,33 +7,15 @@ using Steamworks;
 // - Do not remove achievements or reorder them
 // - Only add them to the bottom of the enum
 // - Otherwise, you will mess up the other references to the enums
-public enum Achievement
-{
-    NONE,
-    MEET_MIKE,
-    MEET_CLAIRE,
-    FINISH_GAME,
-    DRAW_IN_CANVAS,
-    TOUCH_ORANGE_OBJ
-}
 
 public class AchievementUnlocker : MonoBehaviour
 {
-    [SerializeField] private Achievement achievement;
-
-    public static Dictionary<Achievement, string> AchievementIds = new Dictionary<Achievement, string>()
-    {
-        { Achievement.MEET_MIKE, "MEET_MIKE"},
-        { Achievement.MEET_CLAIRE, "MEET_CLAIRE"},
-        { Achievement.FINISH_GAME, "FINISH_GAME"},
-        { Achievement.DRAW_IN_CANVAS, "DRAW_IN_CANVAS"},
-        { Achievement.TOUCH_ORANGE_OBJ, "TOUCH_ORANGE_OBJ"},
-    };
+    [SerializeField] private Global.Achievement achievement;
 
     [ContextMenu("Set Achievement")]
     public void SetAchievement()
     {
-        if (achievement == Achievement.NONE) {
+        if (achievement == Global.Achievement.NONE) {
             Debug.Log("Tried to set achievement but none was specified");
             return;
         }
@@ -44,7 +26,7 @@ public class AchievementUnlocker : MonoBehaviour
         }
 
         // Look up achievement ID
-        if (!AchievementIds.TryGetValue(achievement, out string achievementId)) {
+        if (!Global.AchievementToIdMap.TryGetValue(achievement, out string achievementId)) {
             Debug.Log("Tried to set achievement but could not lookup achievement ID");
             return;
         }
